fix: firebase configuration for app
