+++ src/pages/signup.jsx
@@ -7,10 +7,11 @@ import AuthSocialMedia from "@/components/AuthSocialMedia";
 import FormTitle from "@/components/FormTitle";
 import Button from "@/components/ui/Button";
 import Input from "@/components/ui/Input";
+
 import {
-    signInWithGoogleAccount,
     signInWithFbAccount,
-} from "@/firebase-config/firebaseProvidersMethods";
+    signInWithGoogleAccount,
+} from "@/firebase/firebaseProvidersMethods";
 function SignUp({ t }) {
     const {
         firstname = "firstName",
